requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.3
django-braces==1.16.0
django-crispy-forms==2.3
sqlparse==0.5.2
tzdata==2024.2
